refactor: extract copy-to-clipboard from share_buttons
