+++ foundation_cms/campaigns/templates/campaigns/components/campaign_page/form_state/03_sharing.html
@@ -21,7 +21,7 @@
         </div>
 
         <!-- Copy Link -->
-        <button class="btn btn-secondary btn-secondary--no-arrow btn-secondary--link">
+        <button class="btn btn-secondary btn-secondary--no-arrow btn-secondary--link copy-to-clipboard-button">
             {% trans "Copy Link" %}
         </button>
