+++ foundation_cms/static/js/pages/campaign_page.js
@@ -1,7 +1,7 @@
 import { initConfirmationNoticeBanner } from "../components/campaign_page/confirmation-notice-banner";
-import { initShareButtons } from "../components/campaign_page/share_buttons";
+import { initShareProgressButtons } from "../components/campaign_page/share_buttons";
 import { initNewsletterOptin } from "../components/campaign_page/newsletter_optin";
 
 initConfirmationNoticeBanner();
-initShareButtons();
+initShareProgressButtons();
 initNewsletterOptin();
