+++ foundation_cms/static/js/redesign_main.js
@@ -17,6 +17,7 @@ import { initSpotlightCardCarousels } from "./blocks/spotlight_card_carousel.js"
 import { initPillarCardLinks } from "./blocks/pillar_card_set.js";
 import { initTabbedContent } from "./blocks/tabbed_content_container.js";
 import initBackToTopButton from "./components/back_to_top.js";
+import { initCopyToClipboardButtons } from "./components/copy_to_clipboard.js";
 
 let foundationSiteURL = window.location.origin;
 
@@ -41,3 +42,4 @@ initPillarCardLinks();
 initTabbedContent();
 initBackToTopButton();
 initSearchToggle();
+initCopyToClipboardButtons();
